+++ BE/package.json
@@ -13,6 +13,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@coinfetti/shared-categories": "file:../shared",
     "@supabase/supabase-js": "^2.95.3",
     "dotenv": "^17.2.3",
     "express": "^5.2.1",
